Se añadieron más filtros al CRUD de habitaciones

+++ FinalProject/Models/Opciones.cs
@@ -8,7 +8,7 @@ namespace FinalProject.Models
     public class Opciones
     {
         public int? Sumatoria { get; set; }
-        public int? Conteo { get; set; }
+        public decimal? Conteo { get; set; }
         public decimal? Promedio { get; set; }
         public decimal? MontoMayor { get; set; }
         public decimal? MontoMenor { get; set; }
